+++ src/dependency_injector/__init__.py
@@ -1,6 +1,6 @@
 """Dependency injector top-level package."""
 
-__version__ = '3.10.0'
+__version__ = '3.11.0'
 """Version number that follows semantic versioning.
 
 :type: str